Auto stash before merge of "develop" and "origin/develop"
