+++ BO4E-dotnet/COM/Aufgabe.cs
@@ -2,7 +2,7 @@ using System;
 using BO4E.meta;
 using Newtonsoft.Json;
 using ProtoBuf;
-using static BO4E.BoMapper;
+using static BO4E.meta.LenientConverter.LenientConverter;
 
 namespace BO4E.COM
 {
