+++ BO4ETestProject/BoMapperTests/Vertrag_lenient_String.json
@@ -1,0 +1,139 @@
+{
+  "objectName": "Vertrag",
+  "lenientStringToInt": true,
+  "input": {
+    "versionStruktur": 1,
+    "boTyp": "VERTRAG",
+    "vertragsnummer": "cb32fb41-8bb1-436a-9465-76447b478480",
+    "beschreibung": "Grund- oder Ersatzversorgung",
+    "vertragsart": 0,
+    "vertragstatus": 3,
+    "sparte": 1,
+    "vertragsbeginn": "2020-02-18T00:00:00",
+    "vertragsende": "0001-01-01T00:00:00",
+    "vertragspartner1": {
+      "versionStruktur": 1,
+      "anrede": null,
+      "boTyp": "MARKTTEILNEHMER",
+      "title": null,
+      "name1": "Enercity Gas",
+      "name2": null,
+      "name3": null,
+      "gewerbekennzeichnung": true,
+      "hrnummer": null,
+      "amtsgericht": null,
+      "kontaktweg": null,
+      "umsatzsteuerId": null,
+      "glaeubigerId": null,
+      "eMailAdresse": null,
+      "website": null,
+      "geschaeftspartnerrolle": null,
+      "partneradresse": null,
+      "marktrolle": 0,
+      "rollencodenummer": "9800019400006",
+      "rollencodetyp": 1,
+      "makoadresse": "",
+      "ansprechpartner": null
+    },
+    "vertragspartner2": {
+      "versionStruktur": 1,
+      "anrede": null,
+      "boTyp": "GESCHAEFTSPARTNER",
+      "title": null,
+      "name1": "enercityAG",
+      "name2": null,
+      "name3": "Ersatzversorgung",
+      "gewerbekennzeichnung": true,
+      "hrnummer": null,
+      "amtsgericht": "",
+      "kontaktweg": null,
+      "umsatzsteuerId": null,
+      "glaeubigerId": null,
+      "eMailAdresse": null,
+      "website": null,
+      "geschaeftspartnerrolle": null,
+      "partneradresse": {
+        "postleitzahl": "30167",
+        "ort": "Hannover",
+        "strasse": "Ihmeplatz",
+        "hausnummer": "2",
+        "postfach": null,
+        "adresszusatz": null,
+        "coErgaenzung": null,
+        "landescode": 61
+      }
+    },
+    "unterzeichnervp1": null,
+    "unterzeichnervp2": null,
+    "vertragskonditionen": {
+      "beschreibung": null,
+      "anzahlAbschlaege": "Item12",
+      "vertragslaufzeit": {
+        "einheit": 0,
+        "dauer": -63717580800.0,
+        "startdatum": "2020-02-18T00:00:00",
+        "enddatum": "0001-01-01T00:00:00"
+      },
+      "kuendigungsfrist": null,
+      "vertragsverlaengerung": null,
+      "abschlagszyklus": null,
+      "startAbrechnungsjahr": "0001-01-01T00:00:00",
+      "geplanteTurnusablesung": null,
+      "turnusablesungIntervall": 1,
+      "netznutzungsabrechnung": {
+        "einheit": 4,
+        "dauer": 27.0,
+        "startdatum": "2020-10-01T00:00:00",
+        "enddatum": "2020-10-28T00:00:00"
+      },
+      "netznutzungsabrechnungIntervall": 12,
+      "haushaltskunde": null,
+      "netznutzungsvertrag": 1,
+      "netznutzungszahler": 1,
+      "netznutzungsabrechnungsvariante": 0,
+      "netznutzungsabrechnungsgrundlage": 0
+    },
+    "vertragsteile": [
+      {
+        "vertragsteilbeginn": "2020-04-01T00:00:00",
+        "vertragsteilende": "0001-01-01T00:00:00",
+        "lokation": "NCLB400117640000",
+        "vertraglichFixierteMenge": null,
+        "minimaleAbnahmemenge": null,
+        "maximaleAbnahmemenge": null,
+        "jahresverbrauchsprognose": null,
+        "kundenwert": null,
+        "verbrauchsaufteilung": null
+      }
+    ],
+    "gemeinderabatt": 0.0,
+    "korrespondenzpartner": {
+      "versionStruktur": 1,
+      "anrede": null,
+      "boTyp": "GESCHAEFTSPARTNER",
+      "title": null,
+      "name1": "enercityAG",
+      "name2": null,
+      "name3": "Ersatzversorgung",
+      "gewerbekennzeichnung": true,
+      "hrnummer": null,
+      "amtsgericht": "",
+      "kontaktweg": null,
+      "umsatzsteuerId": null,
+      "glaeubigerId": null,
+      "eMailAdresse": null,
+      "website": null,
+      "geschaeftspartnerrolle": null,
+      "partneradresse": {
+        "postleitzahl": "30167",
+        "ort": "Hannover",
+        "strasse": "Ihmeplatz",
+        "hausnummer": "2",
+        "postfach": null,
+        "adresszusatz": null,
+        "coErgaenzung": null,
+        "landescode": 61
+      }
+    }
+  }
+}
