+++ BO4ETestProject/TestBO4E-dotnet.csproj
@@ -54,6 +54,9 @@
     <None Update="BoMapperTests\energiemenge_profdecimal_verbrauch_bug.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
+    <None Update="BoMapperTests\geschaeftspartner_non_lenient_enumlist.json">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
     <None Update="BoMapperTests\rechnung.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
@@ -87,9 +90,6 @@
     <None Update="BoMapperTests\geschaeftspartner_lenient_enumlist.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
-    <None Update="BoMapperTests\geschaeftspartner_non_lenient_enumlist.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
     <None Update="BoMapperTests\marktlokation_broken_gasqualitaet.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
@@ -100,6 +100,9 @@
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
     <None Update="BoMapperTests\marktlokation_simple.json">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BoMapperTests\Vertrag_lenient_String.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
